Removed UserActivity
The help activity is now the one that opens after selecting the school.  Then the user can select a help category to be redirected to a temporary activity that displays what choices they have made.

NOTE: TestActivity is just a placeholder to demonstrate how to access the UserData that is set when buttons are pressed.  This activity will likely become the FAQ when that gets created.