Added information popups.
Currently the only one with unique data is Keene State -> Local Hospitals
A system has been put into place that already uniquely identifies the combination of buttons.  The only thing that needs to change is the correlated string in the resources file.
Cancel button closes the popup, and Call button does not perform any action yet.