Added the FAQ page
Currently it contains mostly placeholder items, such as a checkbox instead of a button (to see if the answer should be visible) and only one question is legitimate.

Side note: The Fragments seem to be overlapping each other (buttons from previous fragments can still be pressed) so we will have to change the way that we change between them.